# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Personal banking API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /transfers:
    post:
      operationId: createTransfer
      tags: [transfers]
      summary: Initiate a money transfer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TransferInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Transaction"
        "400":
          description: Bad request (insufficient funds, invalid amount, etc.)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /transfers/name-enquiry:
    post:
      operationId: verifyName
      tags: [transfers]
      summary: Verify account holder name from NIBSS EasyPay Name Enquiry
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NameEnquiryInput"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NameEnquiryResponse"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /transfers/{reference}/status:
    get:
      operationId: queryTransferStatus
      tags: [transfers]
      summary: Query NIBSS EasyPay transfer status
      parameters:
        - name: reference
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TransferStatusResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /transfers/webhook/easypay:
    post:
      operationId: easyPayWebhook
      tags: [transfers]
      summary: Webhook for NIBSS EasyPay transfer status callbacks
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EasyPayWebhookInput"
      responses:
        "200":
          description: Callback processed successfully
          content:
            application/json:
              schema:
                type: object
                required: [success]
                properties:
                  success:
                    type: boolean
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    Transaction:
      type: object
      required: [id, accountId, type, amount, currency, description, status, reference, createdAt]
      properties:
        id:
          type: string
        accountId:
          type: string
        type:
          type: string
          enum: [credit, debit]
        amount:
          type: number
        currency:
          type: string
        description:
          type: string
        status:
          type: string
          enum: [pending, completed, failed, reversed]
        reference:
          type: string
        category:
          type: ["string", "null"]
        recipientName:
          type: ["string", "null"]
        recipientBank:
          type: ["string", "null"]
        recipientAccount:
          type: ["string", "null"]
        senderName:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
...
    TransferInput:
      type: object
      required: [fromAccountId, amount, recipientName, recipientBank, recipientAccount, description]
      properties:
        fromAccountId:
          type: string
        amount:
          type: number
          minimum: 1
        recipientName:
          type: string
        recipientBank:
          type: string
        recipientAccount:
          type: string
        description:
          type: string
...
    NameEnquiryInput:
      type: object
      required: [accountNumber, bankCode]
      properties:
        accountNumber:
          type: string
        bankCode:
          type: string
    NameEnquiryResponse:
      type: object
      required: [success]
      properties:
        success:
          type: boolean
        accountName:
          type: string
        error:
          type: string
    TransferStatusResponse:
      type: object
      required: [success, status]
      properties:
        success:
          type: boolean
        status:
          type: string
          enum: [completed, failed, pending]
        error:
          type: string
    EasyPayWebhookInput:
      type: object
      required: [reference, status]
      properties:
        reference:
          type: string
        status:
          type: string
          enum: [completed, failed]
...